+++ operation/benchmarks/opv2/main.py
@@ -68,6 +68,11 @@ def parse_args():
                         required=True,
                         help="result log path for FlagPerf/operation/result")
 
+    parser.add_argument("--image_name",
+                        type=str,
+                        required=True,
+                        help="image name for cp log")
+
     args, unknown_args = parser.parse_known_args()
     args.unknown_args = unknown_args
     return args
@@ -78,10 +83,10 @@ def main(config):
     correctness = correctness == 0
 
     # test operation performance
-    log_dir = do_performance(config.mode, config.warmup, config.log_dir)
-    # performance = performance == 0
-    parse_log_file(config.spectflops, config.mode, config.warmup, log_dir, config.result_log_path)
-    # parse_log_file(config.spectflops, config.mode, config.warmup, config.log_dir, config.result_log_path)
+    performance = do_performance(config.mode, config.warmup, config.log_dir, config.image_name)
+    performance = performance == 0
+    # parse_log_file(config.spectflops, config.mode, config.warmup, config.result_log_path)
+    parse_log_file(config.spectflops, config.mode, config.warmup, config.log_dir, config.result_log_path)
 
     # dtype = {
     #     "FP32": torch.float32,
